+++ packages/react-sdk/src/core/components/CallLayout/SpeakerLayout.tsx
@@ -42,6 +42,11 @@ export type SpeakerLayoutProps = {
    */
   excludeLocalParticipant?: boolean;
   /**
+   * When set to `false` disables mirroring of the local participant's video.
+   * @default true
+   */
+  mirrorLocalParticipantVideo?: boolean;
+  /**
    * Turns on/off the pagination arrows.
    * @default true
    */
@@ -58,6 +63,7 @@ export const SpeakerLayout = ({
   VideoPlaceholder,
   participantsBarPosition = 'bottom',
   participantsBarLimit,
+  mirrorLocalParticipantVideo = true,
   excludeLocalParticipant = false,
   pageArrowsVisible = true,
 }: SpeakerLayoutProps) => {
@@ -117,8 +123,13 @@ export const SpeakerLayout = ({
     );
   }
 
+  const mirror = mirrorLocalParticipantVideo ? undefined : false;
+
   if (!call) return null;
 
+  const renderParticipantsBar =
+    participantsBarPosition &&
+    (participantsWithAppliedLimit.length > 0 || isSpeakerScreenSharing);
   return (
     <div className="str-video__speaker-layout__wrapper">
       <ParticipantsAudio participants={remoteParticipants} />
@@ -134,6 +145,7 @@ export const SpeakerLayout = ({
             <ParticipantView
               participant={participantInSpotlight}
               muteAudio={true}
+              mirror={mirror}
               trackType={
                 isSpeakerScreenSharing ? 'screenShareTrack' : 'videoTrack'
               }
@@ -142,7 +154,7 @@ export const SpeakerLayout = ({
             />
           )}
         </div>
-        {participantsWithAppliedLimit.length > 0 && participantsBarPosition && (
+        {renderParticipantsBar && (
           <div
             ref={setButtonsWrapperElement}
             className="str-video__speaker-layout__participants-bar-buttons-wrapper"
@@ -164,6 +176,7 @@ export const SpeakerLayout = ({
                       participant={participantInSpotlight}
                       ParticipantViewUI={ParticipantViewUIBar}
                       VideoPlaceholder={VideoPlaceholder}
+                      mirror={mirror}
                       muteAudio={true}
                     />
                   </div>
@@ -177,6 +190,7 @@ export const SpeakerLayout = ({
                       participant={participant}
                       ParticipantViewUI={ParticipantViewUIBar}
                       VideoPlaceholder={VideoPlaceholder}
+                      mirror={mirror}
                       muteAudio={true}
                     />
                   </div>