+++ sample-apps/react/react-dogfood/components/Settings/SettingsTabModal.tsx
@@ -30,6 +30,7 @@ import { LanguageMenu } from './LanguageMenu';
 import { CallRecordings } from '../CallRecordings';
 import { useLanguage } from '../../hooks/useLanguage';
 import { useIsProntoEnvironment } from '../../context/AppEnvironmentContext';
+import { IncomingVideoSettingsDropdown } from '../IncomingVideoSettings';
 
 type ToggleSettingsTabModalProps = {
   inMeeting: boolean;
@@ -135,20 +136,23 @@ export const SettingsTabModalMenu = (props: {
   return (
     <SettingsTabModal {...tabModalProps}>
       <TabWrapper icon="device-settings" label={t('Device settings')} inMeeting>
-        <>
-          <DeviceSelectorVideo
-            visualType="dropdown"
-            title={t('Select a Camera')}
-          />
-          <DeviceSelectorAudioInput
-            visualType="dropdown"
-            title={t('Select a Mic')}
-          />
-          <DeviceSelectorAudioOutput
-            visualType="dropdown"
-            title={t('Select a Speaker')}
-          />
-        </>
+        <DeviceSelectorVideo
+          visualType="dropdown"
+          title={t('Select a Camera')}
+        />
+        <DeviceSelectorAudioInput
+          visualType="dropdown"
+          title={t('Select a Mic')}
+        />
+        <DeviceSelectorAudioOutput
+          visualType="dropdown"
+          title={t('Select a Speaker')}
+        />
+        <IncomingVideoSettingsDropdown title={t('Incoming video quality')} />
+        <div className="rd__tab-panel__note">
+          Actual incoming video quality depends on a number of factors, such as
+          the quality of the source video, and network conditions.
+        </div>
       </TabWrapper>
       <TabWrapper icon="video-effects" label="Effects" inMeeting>
         <VideoEffectsSettings />